requirements.txt
paho-mqtt>=1.6.1
pika>=1.2.0
dsnparse>=0.1.15
